+++ CHANGELOG.md
@@ -21,8 +21,9 @@
   * Datei und Klassenname von `consent_manager_theme` geändert in `...\Theme`
   * Datei und Klassenname von `consent_manager_thumbnail_cache` geändert in `...\ThumbnailCache`
   * Datei und Klassenname von `rex_consent_manager_thumbnail_mediamanager` geändert in `...\ThumbnailMediaManager`
+  * Datei und Klassenname von `rex_consent_manager_command_log_delete` geändert in `...\Command\LogDelete`
 * **consent_manager_google_consent_helper:** Datei und Klasse entfernt; nicht mehr in Benutzung 
-
+* **Command log-delete:** aktiviert und verfügbar (in package.yml registriert)
 
 ## Version 4.5.0 - 14.10.2025